requirements.txt
Django==4.2.4
django-cors-headers==4.2.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
mysql-connector==2.2.9
mysqlclient==2.2.0
opencv-python==4.8.1.78
PyMuPDF==1.23.6
pytesseract==0.3.10
tesseract==0.1.3
gunicorn ==20.0.4
django-sslserver==0.22